Disable the formatter around dicts in the part import template
Because we're using this file as a format string, any curly braces that we want in the output (eg. dictionaries) need to be doubled up or they will be interpreted as a single pair of curly braces (ie will try to be handled by the .format() call). Doubling them up causes the formatter to freak out a little and split them onto separate lines so just switch off the formatter around dict definitions.